requirements.txt
numpy==1.22.0
opencv-python==4.5.5.62
Pillow==9.0.0
rdkit-pypi==2021.9.4
